Small changes to the zero point values.
The "VEGAMAG" zeropoints for NIRISS and Guider have changed slightly due to a recommendation that Sirius be used as the A0V standard template spectrum with magnitude -1.401 (the previous value was taken to be -1.415).  In regenerating the file with code, it was found that the Guider F-lambda zero point values were off by a factor of 10 (although the F_nu ones were not).